Add StoryImageView and some

+++ Stock/Models/NewsModel.swift
@@ -9,7 +9,7 @@ import Foundation
 
 struct NewsModel: Codable {
     let category: String
-    let datetime: Int
+    let datetime: Double
     let headline: String
     let id: Int
     let image: String
